added kraken api, and removed it from cryptowatch
Unfortunately kraken doesn't really provide an overall market summery with price info as many other APIs do. This makes it necessary to include all the desired coinpairs in the URL. The "safest" way to do this would probably be to have a function in the kraken markets object that gets all the kraken asset pairs from the API, parses them out of the JSON, and then adds them to a URL and returns that. But that seems like a lot of code to write, and would only have the advantage of never having to worry about adding a coin pair manually when kraken adds one to the exchange - which may happen, maybe once or twice a year, tops?